✨ feat: add dialog to prompt company configuration in Sidebar component
